fix(migration-engine): changed type of "compositeTypeDepth" from u32 to isize, fixing type mismatch issues with the TS cli

+++ migration-engine/core/src/state.rs
@@ -322,7 +322,7 @@ impl GenericApi for EngineState {
             &params.schema,
             None,
             Box::new(move |connector| {
-                let composite_type_depth = From::from(params.composite_type_depth as isize);
+                let composite_type_depth = From::from(params.composite_type_depth);
                 let ctx = migration_connector::IntrospectionContext::new(schema, composite_type_depth);
                 Box::pin(async move {
                     // TODO(MultiSchema): Grab namespaces from introspect params?
